[#36] refactor : UserCoupon 위치 이동
- user Coupon 위치 이동 및
- wish entity 수정
